Refactor Docker client to add Windows support
